arch: split config module to reduce fan-in from 17 to 12 (#172)
Move domain types out of config.rs into their proper domain modules:
- ModelDef/TransitionDef → domain::statemachine (fix 3 stale imports)
- SessionMode/AgentRuntime → domain::agent (new module)

Re-exports in config.rs preserve backward compatibility.
Config fan-in reduced from 17 → 12 (target was ≤15).

Closes #172
